requirements 2.txt
torch
# Discord AI Bot dependencies
discord.py
fastapi
uvicorn
openai
pinecone-client
python-dotenv
sentence-transformers
unstructured
spacy 